+++ R/Data.R
@@ -42,6 +42,7 @@ NULL
 #' @param GroupingVariables The variables to report by. For most applications \code{GroupingVariables} should include "Survey" and "SpeciesCategory", unless the user needs to sum over all Survey or SpeciesCategory.
 #' @param InformationVariables Variables to include as columns to the end of the report table. These cannot have more unique combinations than the \code{GroupingVariables}.
 #' @param RemoveMissingValues Logical: If TRUE, remove missing values (NAs) from the \code{TargetVariable}. The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA. Use \code{RemoveMissingValues} = TRUE with extreme caution, as it may lead to under-estimation. E.g., if \code{RemoveMissingValues} = TRUE and a super-individual lacks \code{IndividualRoundWeight}, \code{Biomass} will be NA, and the portion of \code{Abundance} distributed to that super-individual will be excluded when summing \code{Biomass} (but included when summing \code{Abundance}). It is advised to always run with \code{RemoveMissingValues} = FALSE first, and make a thorough investigation to identify the source of any missing values. The function \code{link{ImputeSuperIndividuals}} can be used to impute the missing information from other super-individuals.
+#' @param Filter A string with an R expression to filter out unwanted rows of the report, e.g. "IndividualAge \%notin\% NA" or "Survey \%notin\% NA & SpeciesCategory \%notin\% NA".
 #' 
 #' @name general_report_arguments
 #' 
@@ -678,15 +679,15 @@ NULL
 NULL
 
 
-# ##################################################
-# ##################################################
-# #' Write StratumPolygon data
-# #' 
-# #' The WriteStratumPolygonData model data is used to write \code{\link{StratumPolygon}}.
-# #' 
-# #' @seealso This data type is produced by \code{\link{WriteStratumPolygon}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}
-# #' 
-# #' @name WriteStratumPolygonData
-# #' 
-# NULL
+##################################################
+##################################################
+#' Write StratumPolygon data
+#' 
+#' The WriteStratumPolygonData model data is used to write \code{\link{StratumPolygon}}.
+#' 
+#' @seealso This data type is produced by \code{\link{WriteStratumPolygon}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}
+#' 
+#' @name WriteStratumPolygonData
+#' 
+NULL